backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "multimodal-ai-backend"
version = "0.1.0"
description = "Personal Multimodal AI Assistant Backend with Screen Context"
readme = "MULTIMODAL_README.md"
requires-python = ">=3.11"
authors = [
    {name = "Khushal", email = "your-email@example.com"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core Framework (CPU-only for development)
    "torch>=2.1.0",
    "torchvision>=0.16.0", 
    "torchaudio>=2.1.0",
    
    # Primary Cloud Services
    "openai>=1.3.0",
    "huggingface-hub>=0.19.0",
    "google-generativeai>=0.3.0",
    
    # Essential ML Libraries
    "transformers>=4.35.0",
    "datasets>=2.14.0",
    "tokenizers>=0.15.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-google-genai>=0.0.6",
    
    # FastAPI Backend
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "websockets>=12.0",
    "python-multipart>=0.0.6",
    "httpx>=0.25.0",
    
    # Audio Processing (WebRTC)
    "pyaudio>=0.2.11",
    "pydub>=0.25.0",
    "librosa>=0.10.1",
    "soundfile>=0.12.0",
    "numba>=0.56.0",
    
    # Data Processing
    "numpy>=1.24.0",
    "pandas>=2.1.0",
    "Pillow>=10.0.0",
    
    # Validation and Configuration
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    
    # Code Quality and Environment
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    # Visualization
    "matplotlib>=3.7.0",
    "plotly>=5.17.0",
    "streamlit>=1.28.0",
    
    # Code Quality
    "black>=23.9.0",
    "isort>=5.12.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

anthropic = [
    "anthropic>=0.7.0",
]

aws = [
    "boto3>=1.34.0",
]

[project.urls]
Homepage = "https://github.com/khushal/background-multimodal-llm"
Repository = "https://github.com/khushal/background-multimodal-llm"

[tool.black]
target-version = ['py311']
line-length = 200

[tool.isort]
multi_line_output = 3
known_first_party = ["multimodal_ai_backend"]
line_length = 200

[tool.flake8]
max-line-length = 200
extend-ignore = ["E203", "W503"]

[tool.pylint]
max-line-length = 200

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v"
asyncio_mode = "auto" 